# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API to get lists, store, update and delete change logs, companies, projects and categories",
        "version": "0.1"
    },
    "paths": {
        "/api/{company}/categories": {
            "get": {
                "summary": "list of company categories",
                "operationId": "App\\Http\\Controllers\\CategoryController::getByCompanyId",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/project/{projectUuid}/changelogs": {
            "post": {
                "summary": "store project new changelog",
                "operationId": "App\\Http\\Controllers\\ChangelogController::store",
                "parameters": [
                    {
                        "name": "projectUuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_at",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/project/changelogs/{id}": {
            "put": {
                "summary": "update project changelog",
                "operationId": "App\\Http\\Controllers\\ChangelogController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_at",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "delete": {
                "summary": "delete specific project changelog",
                "operationId": "App\\Http\\Controllers\\ChangelogController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/{project}/changelogs": {
            "get": {
                "summary": "list of project changelogs",
                "operationId": "App\\Http\\Controllers\\ProjectController::getChangeLogs",
                "parameters": [
                    {
                        "name": "projectUuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/{project_uuid}/published/changelogs": {
            "get": {
                "summary": "list of project's published changelogs",
                "operationId": "App\\Http\\Controllers\\ProjectController::getPublishedChangelogs",
                "parameters": [
                    {
                        "name": "projectUuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "uuid": {
                "description": "A unique identififer for the uploaded file.",
                "type": "char"
            }
        }
    },
    "security": [
        []
    ]
}